Basket route fix

+++ db/migrate/20171212141550_create_baskets.rb
@@ -3,7 +3,7 @@ class CreateBaskets < ActiveRecord::Migration[5.0]
     create_table :baskets do |t|
       t.references :user, foreign_key: true
       t.references :truck_order_list, foreign_key: true
-      t.integer :total_price
+      t.float :total_price
       t.date :date
       t.datetime :time
       t.string :status
